refactor: create structs and constants
Also export all values (interfaces, enums, types and constants) from them
