+++ src/constants.ts
@@ -1,0 +1,6 @@
+/**
+ * Field on DOM node that stores the previous VNode
+ */
+const OLD_VNODE_FIELD = '__m_old_vnode';
+
+export { OLD_VNODE_FIELD }
