+++ src/m.ts
@@ -1,19 +1,4 @@
-export type VProps = Record<string, string | unknown | (() => void)>;
-export type VNode = VElement | string;
-
-export interface VElement {
-  tag: string;
-  props?: VProps;
-  children?: VNode[];
-  key?: string;
-  flag?: VFlags;
-}
-
-export enum VFlags {
-  NO_CHILDREN = 0,
-  ONLY_TEXT_CHILDREN = 1,
-  ANY_CHILDREN = 2,
-}
+import { VElement, VFlags, VNode, VProps } from "./structs";
 
 /**
  * Attaches ns props to svg element
